+++ pkg/manager/diff/patch_test.go
@@ -1,7 +1,7 @@
-// Copyright 2025 syzkaller project authors. All rights reserved.
+// Copyright 2026 syzkaller project authors. All rights reserved.
 // Use of this source code is governed by Apache 2 LICENSE that can be found in the LICENSE file.
 
-package manager
+package diff
 
 import (
 	"fmt"
@@ -106,17 +106,4 @@ func TestModifiedSymbols(t *testing.T) {
 		patched["function2"] = "hash2"
 		assert.Equal(t, []string{"function", "function2"}, modifiedSymbols(base, patched))
 	})
-}
-
-func TestNeedReproForTitle(t *testing.T) {
-	for title, skip := range map[string]bool{
-		"no output from test machine":                          false,
-		"SYZFAIL: read failed":                                 false,
-		"lost connection to test machine":                      false,
-		"INFO: rcu detected stall in clone":                    false,
-		"WARNING in arch_install_hw_breakpoint":                true,
-		"KASAN: slab-out-of-bounds Write in __bpf_get_stackid": true,
-	} {
-		assert.Equal(t, skip, needReproForTitle(title), "title=%q", title)
-	}
 }
